requirements.txt
pandas
requests
python-dotenv
google-cloud-bigquery
pyarrow